+++ inc/form_element_radio.js
@@ -23,6 +23,15 @@ form_element_radio.prototype.connect=function(dom_parent) {
   }
 }
 
+form_element_radio.prototype.focus = function() {
+  let list = Object.keys(this.dom_values)
+  if (list.length === 0) {
+    return
+  }
+
+  this.dom_values[list[0]].focus()
+}
+
 form_element_radio.prototype.get_data=function() {
   if(!this.dom_values)
     return this.data;